infra_platform/requirements.txt
requests
tenacity
kafka-python